Fix paths
