+++ MeasureWatchFace/.run/Watchface.run.xml
@@ -1,0 +1,20 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Watchface" type="AndroidWatchFaceConfigurationType" factoryName="AndroidWatchFaceConfigurationFactory" show_console_on_std_err="false" show_console_on_std_out="false">
+    <module name="MeasureWatchFace.app" />
+    <option name="allowRunningInParallel" value="false" />
+    <option name="projectPathOnTarget" />
+    <option name="selectedOptions">
+      <list />
+    </option>
+    <LaunchOptions>
+      <option name="componentName" value="com.garan.measurewatchface.MeasureWatchFaceService" />
+    </LaunchOptions>
+    <DeployOptions>
+      <option name="pmInstallFlags" value="" />
+    </DeployOptions>
+    <method v="2">
+      <option name="Make" enabled="true" />
+      <option name="Android.Gradle.BeforeRunTask" enabled="true" />
+    </method>
+  </configuration>
+</component>
