+++ MeasureWatchFace/watchface-benchmark/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.google.android.horologist.media.benchmark">
+    package="com.google.android.horologist.benchmark">
 
     <uses-feature android:name="android.hardware.type.watch" />
