+++ MeasureWatchFace/watchface-benchmark/src/main/java/com/google/android/horologist/benchmark/WatchfaceActions.kt
@@ -1,4 +1,4 @@
-package com.google.android.horologist.media.benchmark
+package com.google.android.horologist.benchmark
 
 import androidx.benchmark.macro.MacrobenchmarkScope
